examples/web/requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
pyvis>=0.3.2